requirements.txt
llama_index
llama-index-llms-gemini
python-dotenv==1.0.1
rich==13.7.0
